pva: mirror from gitlab cv/pva-sys-sw
Gitlab commit b284345610ceb ("pva: fix static defects...")

No changes since last deployment.

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3428161
Tested-by: Rohit Goel <rogoel@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/host/pva/src/kmd/common/shim/pva_kmd_shim_ccq.h
@@ -14,7 +14,8 @@ struct pva_kmd_device;
  *
  */
 void pva_kmd_ccq_push(struct pva_kmd_device *pva, uint8_t ccq_id,
-		      uint64_t ccq_entry);
+		      uint32_t ccq_entry);
+
 /**
  * @brief Get the number of available spaces in the CCQ.
  *
